requirements.txt
boto3>=1.26.0
python-dotenv>=0.19.0
requests>=2.28.0 